fix up tailwind docs
