+++ docs/Systems/Tailwind/utility-class-helpers/backgrounds/background-color.mdx
