+++ docs/Systems/Tailwind/utility-class-helpers/borders/border-color.mdx
