+++ docs/Systems/Tailwind/utility-class-helpers/borders/border-color.stories.tsx
@@ -3,42 +3,40 @@ import { Meta, StoryFn } from "@storybook/react"
 import classnames from "classnames"
 import { kaizenTailwindTheme } from "@kaizen/tailwind"
 import { UtilityClassTemplate } from "../../components/UtilityClassTemplate"
+import { flattenEntries } from "../../helpers/flattenEntries"
 import { utilityDescription } from "../../helpers/utilityDescription"
 
 const prefix = "border-"
-const classEntries: Array<{ utilityClassName: string; cssProperty: string }> =
-  Object.entries(kaizenTailwindTheme?.borderWidth || []).map(
-    ([suffix, cssProperty]) => ({
-      utilityClassName: `${prefix}${suffix}`,
-      cssProperty,
-    })
-  )
+const classEntries = flattenEntries(
+  prefix,
+  kaizenTailwindTheme?.borderColor || {}
+)
 
 export default {
-  title: "Systems/Tailwind/Utility Class References/Borders/Border Width",
+  title: "Systems/Tailwind/Utility Class References/Borders/Border Color",
   parameters: {
     a11y: { disable: true },
     chromatic: { disable: false },
     docsLayout: "fullPage",
     docs: {
       description: {
-        component: utilityDescription(prefix, classEntries[1].utilityClassName),
+        component: utilityDescription(prefix, classEntries[0].utilityClassName),
       },
     },
   },
 } satisfies Meta
 
-export const BorderWidth: StoryFn<{ isReversed: boolean }> = ({
+export const BorderColor: StoryFn<{ isReversed: boolean }> = ({
   isReversed,
 }) => (
   <UtilityClassTemplate
-    compiledCssPropertyName="border-width"
+    compiledCssPropertyName="border-color"
     classKeyValues={classEntries}
     renderExampleComponent={(utilityClass): React.ReactElement => (
       <div
         className={classnames(
-          "w-[100px] h-[100px] border-[black] border",
-          utilityClass
+          "rounded w-[100px] h-[100px] border",
+          utilityClass.replace("-DEFAULT", "")
         )}
       />
     )}
