+++ docs/Systems/Tailwind/utility-class-helpers/typography/font-weight.mdx
