+++ docs/Systems/Tailwind/utility-class-helpers/typography/max-text-width.mdx
