+++ docs/Systems/Tailwind/utility-class-helpers/typography/text-color.mdx
